Add Conforma Workflow high level diagram
This commit adds a new section in the user guide, named Diagrams. The
purpose of this section is to collect several diagrams to capture the
high-level workflows and architecture of the Conforma project.

The first diagram outlines the workflow of Conforma within the Konflux
pipelines, with detailed descriptions for each step.

Assisted by: Claude Code

Ref: https://issues.redhat.com/browse/EC-1270